refactor: Splitted renderer HTML endpoints into separate files
