+++ src/preload/docs.ts
