+++ src/preload/options.ts
