+++ src/renderer/offline/index.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+	<meta charset="UTF-8">
+	<meta name="viewport" content="width=device-width, initial-scale=1.0">
+	<title>Offline</title>
+	<link rel="stylesheet" href="./style.css" />
+</head>
+
+<body>
+	<h1>
+		<picture>
+			<source media="(prefers-color-scheme: dark)" srcset="../icons/dark/offline.svg">
+			<source media="(prefers-color-scheme: light)" srcset="../icons/light/offline.svg">
+			<img alt="" src="../icons/dark/offline.svg">
+		</picture>&emsp;You are offline
+	</h1>
+	<p>Please check your Internet connection and try again.</p>
+	<button id="retry" type="button">Retry</button>
+	<p id="loading">Loading...</p>
+
+	<script type="module" src="./script.ts"></script>
+</body>
+
+</html>
