+++ src/renderer/offline/style.css
@@ -1,10 +1,3 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<meta charset="UTF-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<title>Offline</title>
-	<style>
 :root {
 	--bg: #191919;
 	--bg-sidebar: #202020;
@@ -18,6 +11,7 @@
 	--font-family: ui-sans-serif, -apple-system, BlinkMacSystemFont, "Segoe UI Variable Display", "Segoe UI", Helvetica, "Apple Color Emoji", Arial, sans-serif, "Segoe UI Emoji", "Segoe UI Symbol";
 	--font: 500 14px/22px var(--font-family);
 }
+
 @media (prefers-color-scheme: light) {
 	:root {
 		--bg: #fff;
@@ -30,9 +24,12 @@
 		--color-hover: #1d1b16;
 	}
 }
-html, body {
+
+html,
+body {
 	background-color: var(--bg);
 }
+
 body {
 	display: flex;
 	flex-direction: column;
@@ -41,20 +38,27 @@ body {
 	font: var(--font);
 	gap: 10px;
 }
-h1, p, button {
+
+h1,
+p,
+button {
 	padding: 0;
 	margin: 0;
 }
+
 h1 {
 	color: var(--color-hover);
 }
+
 img {
 	aspect-ratio: 1;
 	height: 23px;
 }
+
 p {
 	margin-left: 40px;
 }
+
 #retry {
 	display: block;
 	width: fit-content;
@@ -65,25 +69,12 @@ p {
 	color: currentColor;
 	margin: 20px 40px;
 }
+
 #retry:hover {
 	border-color: var(--border-sidebar-hover);
 	background-color: var(--bg-selected);
 }
+
 #loading {
 	display: none;
 }
-	</style>
-</head>
-<body>
-	<h1><picture>
-		<source media="(prefers-color-scheme: dark)" srcset="./icons/dark/offline.svg">
-		<source media="(prefers-color-scheme: light)" srcset="./icons/light/offline.svg">
-		<img alt="" src="./icons/dark/offline.svg">
-	</picture>&emsp;You are offline</h1>
-	<p>Please check your Internet connection and try again.</p>
-	<button id="retry" type="button">Retry</button>
-	<p id="loading">Loading...</p>
-
-	<script type="module" src="./offline.ts"></script>
-</body>
-</html>
