+++ src/renderer/titlebar/style.css
@@ -1,10 +1,3 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<meta charset="UTF-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<title>Notion Electron</title>
-	<style>
 :root {
 	--bg: #191919;
 	--bg-sidebar: #202020;
@@ -18,6 +11,7 @@
 	--font-family: ui-sans-serif, -apple-system, BlinkMacSystemFont, "Segoe UI Variable Display", "Segoe UI", Helvetica, "Apple Color Emoji", Arial, sans-serif, "Segoe UI Emoji", "Segoe UI Symbol";
 	--font: 500 14px/22px var(--font-family);
 }
+
 @media (prefers-color-scheme: light) {
 	:root {
 		--bg: #fff;
@@ -30,6 +24,7 @@
 		--color-hover: #1d1b16;
 	}
 }
+
 body {
 	app-region: drag;
 	user-select: none;
@@ -55,6 +50,7 @@ body {
 	display: flex;
 	flex: 0 0 108px;
 }
+
 .control {
 	display: flex;
 	flex: 0 0 36px;
@@ -71,12 +67,15 @@ body {
 	padding: 0;
 	box-sizing: border-box;
 }
+
 .control:hover {
 	opacity: 1;
 }
+
 .control:disabled {
 	opacity: 0.1;
 }
+
 .control picture {
 	width: 24px;
 	height: 24px;
@@ -91,12 +90,14 @@ body {
 	overflow-x: scroll;
 	scrollbar-width: none;
 }
+
 .tabgroup-container {
 	flex: 1 1 auto;
 	width: 100%;
 	position: relative;
 	overflow: hidden;
 }
+
 .tabgroup-shadow-right {
 	width: 20px;
 	height: 40px;
@@ -105,6 +106,7 @@ body {
 	right: 0;
 	top: 0;
 }
+
 .tabgroup-shadow-left {
 	width: 20px;
 	height: 40px;
@@ -144,22 +146,27 @@ body {
 	position: relative;
 	app-region: no-drag;
 }
+
 .tab.selected,
 .tab:hover {
 	background-color: var(--bg);
 	color: var(--color-hover);
 }
+
 .tab:hover .tab-close {
 	visibility: visible;
 }
+
 .tab.right {
 	margin-left: auto;
 }
+
 .tab-title {
 	overflow: hidden;
 	white-space: nowrap;
 	text-overflow: ellipsis;
 }
+
 .tab-close {
 	display: inline-block;
 	width: 22px;
@@ -173,12 +180,15 @@ body {
 	color: var(--color);
 	app-region: no-drag;
 }
+
 .tab-close:hover {
 	color: var(--color-hover);
 }
+
 .tabgroup-pinned .tab {
 	min-width: 40px;
 }
+
 .tabgroup-pinned .tab::before {
 	content: '';
 	position: absolute;
@@ -189,25 +199,31 @@ body {
 	border-radius: 4px;
 	background-color: var(--border-sidebar);
 }
+
 .tabgroup-pinned .tab.selected::before,
 .tabgroup-pinned .tab:hover::before {
 	background-color: var(--color);
 }
+
 .tabgroup-pinned .tab .tab-close,
 .tabgroup-pinned .tab .tab-title {
 	display: none;
 }
+
 .tab picture,
 .tab-icon {
 	width: 16px;
 	height: 16px;
 }
+
 .tab picture {
 	padding-left: 4px;
 }
+
 .tab.sortable-ghost {
 	visibility: hidden;
 }
+
 .tab.sortable-ghost:hover .tab-close {
 	visibility: hidden;
 }
@@ -233,6 +249,7 @@ body {
 	justify-content: end;
 	padding: 0 5px;
 }
+
 .sidebar.collapsed {
 	flex: 0 0 30px;
 	border: none;
@@ -240,21 +257,27 @@ body {
 	justify-content: center;
 	min-width: 30px;
 }
+
 .sidebar-collapsed {
 	display: none;
 }
+
 .sidebar.collapsed .sidebar-collapsed {
 	display: flex;
 }
+
 .sidebar.collapsed .sidebar-unfold {
 	display: none;
 }
+
 .sidebar-collapsed .sidebar-icon-unfold {
 	display: none;
 }
+
 .sidebar-collapsed:hover .sidebar-icon-menu {
 	display: none;
 }
+
 .sidebar-collapsed:hover .sidebar-icon-unfold {
 	display: inline-block;
 }
@@ -263,13 +286,16 @@ body {
 	.titlebar {
 		background-color: var(--bg-sidebar);
 	}
+
 	.tab:hover,
 	.tab.selected {
 		background-color: var(--bg);
 	}
+
 	.tabgroup-shadow-right {
 		background: linear-gradient(to right, transparent, var(--bg-sidebar));
 	}
+
 	.tabgroup-shadow-left {
 		background: linear-gradient(to left, transparent, var(--bg-sidebar));
 	}
@@ -278,94 +304,3 @@ body {
 .hidden {
 	display: none;
 }
-	</style>
-</head>
-<body>
-	<template id="tab-template">
-		<button class="tab" role="button">
-			<picture>
-				<source class="tab-icon-source" srcset="./icons/light/document.svg" media="(prefers-color-scheme: light)" />
-				<img class="tab-icon" src="./icons/dark/document.svg" />
-			</picture>
-			<span class="tab-title">Tab</span>
-			<span class="tab-close">×</span>
-		</button>
-	</template>
-	<div class="titlebar">
-		<div class="sidebar collapsed hidden">
-			<button class="control sidebar-collapsed" role="button" title="Open Sidebar">
-				<picture class="sidebar-icon-menu">
-					<source srcset="./icons/light/menu.svg" media="(prefers-color-scheme: light)" />
-					<img class="icon" src="./icons/dark/menu.svg" />
-				</picture>
-				<picture class="sidebar-icon-unfold">
-					<source srcset="./icons/light/double_arrow_right.svg" media="(prefers-color-scheme: light)" />
-					<img class="icon" src="./icons/dark/double_arrow_right.svg" />
-				</picture>
-			</button>
-			<button class="control sidebar-unfold" role="button" title="Collapse Sidebar">
-				<picture>
-					<source srcset="./icons/light/double_arrow_left.svg" media="(prefers-color-scheme: light)" />
-					<img class="icon" src="./icons/dark/double_arrow_left.svg" />
-				</picture>
-			</button>
-		</div>
-		<div class="controls">
-			<button
-				class="control"
-				role="button"
-				title="Back"
-				data-action="history-back"
-				disabled
-			>
-				<picture>
-					<source srcset="./icons/light/arrow_left.svg" media="(prefers-color-scheme: light)" />
-					<img class="icon" src="./icons/dark/arrow_left.svg" />
-				</picture>
-			</button>
-			<button
-				class="control"
-				role="button"
-				title="Forward"
-				data-action="history-forward"
-				disabled
-			>
-				<picture>
-					<source srcset="./icons/light/arrow_right.svg" media="(prefers-color-scheme: light)" />
-					<img class="icon" src="./icons/dark/arrow_right.svg" />
-				</picture>
-			</button>
-			<button
-				class="control"
-				role="button"
-				title="Open New Tab"
-				data-action="add"
-			>
-				<picture>
-					<source srcset="./icons/light/add.svg" media="(prefers-color-scheme: light)" />
-					<img class="icon" src="./icons/dark/add.svg" />
-				</picture>
-			</button>
-		</div>
-
-		<div class="tabgroup-container">
-			<div class="tabgroup-shadow-left hidden"></div>
-			<div class="tabgroup"></div>
-			<div class="tabgroup-shadow-right hidden"></div>
-		</div>
-
-		<div class="tabgroup-pinned"></div>
-
-		<div class="tab-all">
-			<button class="control" role="button" title="All Tabs">
-				<picture>
-					<source srcset="./icons/light/arrow_down.svg" media="(prefers-color-scheme: light)" />
-					<img class="icon" src="./icons/dark/arrow_down.svg" />
-				</picture>
-			</button>
-		</div>
-	</div>
-
-	<script type="module" src="./titlebar.ts"></script>
-</body>
-</html>
